backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pandas==2.1.4
openpyxl==3.1.2
pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiofiles==23.2.1
python-barcode==0.15.1
# 모바일 웹 자동화 (Playwright - Chrome mobile emulation)
playwright==1.40.0
# Android 네이티브 앱 자동화 (Appium)
Appium-Python-Client==3.1.1
